Vignette compare sim with ref (#6)
* Make daisy_time_to_timestamp work with list of dlfs + use POSIXct

ggplot needs POSIXct instead of POSIXlt. So we change it here to
avoid having to deal with it everytime we plot

* Rewrite plot_many

We no longer use reshape. This has two benefits
1. We can have multiple y values for the same x value
2. We no longer need to convert x values of POSIXt, so we can drop
the parameter x_var_is_posix_time

* Generic point_and_lines plot function

* Linting

* Vignette showing how to plot field and simulated data together

point_and_lines now only uses cowplot::plot_grid when there are
multiple y_vars. This allows the caller to add data to the plot
when plotting a single variable.